remove readings.h and .c
