+++ sources/model/outbound_message_factory.c
@@ -17,7 +17,6 @@
 #include "outbound_message_factory.h"
 #include "model/feed.h"
 #include "model/outbound_message.h"
-#include "model/reading.h"
 #include "protocol/parser.h"
 #include "utility/wolk_utils.h"
 
@@ -43,8 +42,8 @@ size_t outbound_message_make_from_feeds(parser_t* parser, const char* device_key
 
     parser->create_topic(parser->D2P_TOPIC, device_key, parser->FEED_VALUES_MESSAGE_TOPIC, topic);
 
-    num_serialized = parser_serialize_feeds(parser, readings, type, readings_number, reading_element_size, payload,
-                                               sizeof(payload));
+    num_serialized =
+        parser_serialize_feeds(parser, readings, type, readings_number, reading_element_size, payload, sizeof(payload));
     if (num_serialized != 0)
         outbound_message_init(outbound_message, topic, payload);
