+++ sources/protocol/parser.c
@@ -93,7 +93,7 @@ void parser_init(parser_t* parser)
 }
 
 size_t parser_serialize_feeds(parser_t* parser, feed_t* readings, data_type_t type, size_t num_readings,
-                                 size_t reading_element_size, char* buffer, size_t buffer_size)
+                              size_t reading_element_size, char* buffer, size_t buffer_size)
 {
     /* Sanity check */
     WOLK_ASSERT(parser);
